Introduce a new FunWithFlags.Supervisor module, which is a module-based Supervisor.
Managing the supervison tree for the package is moved there, and the Application simply delegates to it. This commit does not change the supervision tree nor its strategy.
